refactor: share components and integrate stats in popup
