+++ components/StatsTab.js
@@ -1,6 +1,6 @@
-import { h } from './../../js/lib/preact.mjs';
-import htm from './../../js/lib/htm.mjs';
-import { getMessage } from './../../js/modules/i18n.js';
+import { h } from './../js/lib/preact.mjs';
+import htm from './../js/lib/htm.mjs';
+import { getMessage } from './../js/modules/i18n.js';
 
 const html = htm.bind(h);
